MDL-82156 core: Remove manually loaded classes
These classes were all introduced in Moodle 4.4.

Originally they were scheduled for removal in MDL-81134, but now that we
have MDL-82156 we can remove them immediately as the autoloader now
checks for the presence of a newer class and resets if that class is not
present in the autoloader cache.