MDL-83083 theme: Revert preview icon
In MDL-82211, the preview icon was updated to use the "fa-regular fa-eye" icon.
However, this icon is also used elsewhere to represent hide/show functionality,
leading to potential confusion.
To address this, the UX team has requested that the change be reverted, and the
preview icon should now be replaced with "fa-magnifying-glass-plus".