MDL-83092 AI: Placement action needs to check status of provider action
When all providers for a certain action are disabled, or if providers
are enabled but all individual actions of a type are disabled, then
a warning is shown on the placement admin settings page for actions
that have no providers available.

Also the text editor placement will not show the option to generate
images or text if there are no actions available.