MDL-82151 core: Ensure that upgrade notes are sorted
Prior to this change the order of upgrade notes changed depending on
what was encountered first when looping through the list of upgrade
notes.

This change ensures that the core component is always first, followed by
core subsystems sorted lexically, and finally a lexical sorting of
everything else.

The note types are also sorted in the order suggested at keep a
changelog, that is:
- Added
- Changed
- Deprecated
- Removed
- Fixed